remove unnecessary cast
git-svn-id: https://repo.algotrader.ch/svn/trunk/AlgoTrader@3101 19eef36c-86f0-de11-afab-0004ac968479

+++ src/com/algoTrader/entity/PositionImpl.java
@@ -21,7 +21,7 @@ public class PositionImpl extends com.algoTrader.entity.Position {
     public double getValueDouble() {
 
         if (isOpen()) {
-            return (double)getQuantity() * getSecurity().getCurrentValuePerContractDouble();
+            return getQuantity() * getSecurity().getCurrentValuePerContractDouble();
         } else {
             return 0.0;
         }
